Refactor: 콘텐츠 블록 컴포넌트 분리 및 유틸함수 정리
