+++ src/app/admin/notice/create/page.tsx
@@ -3,27 +3,59 @@
 import { useRouter } from 'next/navigation';
 import { BaseSyntheticEvent } from 'react';
 import { useMutation } from '@tanstack/react-query';
-import { FormProvider, useFieldArray, useForm } from 'react-hook-form';
+import { FormProvider, useForm } from 'react-hook-form';
 
 import * as styles from './page.css';
 
 import createNotice from '@/app/_api/notice/createNotice';
 import uploadNoticeImages from '@/app/_api/notice/uploadNoticeImages';
 
-import { NOTICE_CONTENT } from '@/lib/constants/notice';
-import { ItemsType, NoticeContentsType, NoticeCreateType } from '@/lib/types/noticeType';
+import { NoticeCreateType } from '@/lib/types/noticeType';
+import { noticeDescriptionRules, noticeTitleRules } from '@/lib/constants/formInputValidationRules';
 
-import ContentsContainer from './_components/ContentsContainer';
 import CategoryDropdown from './_components/CategoryDropdown';
+import ContentsBody from './_components/ContentsBody';
 
-const noticeTitleRules = {
-  required: '제목은 필수값입니다.',
-  maxLength: { value: 30, message: '제목은 최대 30자까지 입력할 수 있어요.' },
+/** 공지 작성 데이터 포맷 유틸 함수 */
+const formatNoticeData = (originData: NoticeCreateType) => {
+  // order 정리 및 이미지 url 초기화
+  const updatedContents = originData.contents.map((item, index) => {
+    const newContents = { ...item, order: index + 1 };
+    if (newContents.type === 'image') {
+      newContents.imageUrl = '';
+    }
+    return newContents;
+  });
+
+  // 새로운 데이터 객체 반환
+  const noticeData: NoticeCreateType = {
+    ...originData,
+    contents: updatedContents,
+  };
+  return noticeData;
 };
 
-const noticeDescriptionRules = {
-  required: '소개는 필수값입니다.',
-  maxLength: { value: 30, message: '소개는 최대 30자까지 입력할 수 있어요.' },
+/** 이미지 업로드 데이터 포맷 유틸 함수 */
+const formatImageData = (originData: NoticeCreateType) => {
+  const imageExtensionData = originData.contents
+    .map((item, index) => {
+      return { ...item, order: index + 1 };
+    })
+    .filter((item) => item.type === 'image' && item.imageUrl !== '')
+    .map(({ order, imageUrl }) => {
+      return {
+        order: order,
+        extension: (imageUrl as File).type.split('/')[1],
+      };
+    });
+
+  const imageFileData = originData.contents
+    .filter((item) => item.type === 'image' && item.imageUrl !== '')
+    .map((item) => {
+      return item.imageUrl as File;
+    });
+
+  return { imageExtensionData, imageFileData };
 };
 
 export default function CreateNotice() {
@@ -45,88 +77,7 @@ export default function CreateNotice() {
   });
 
   // prettier-ignore
-  const { register, handleSubmit, control, formState: { errors, isValid }} = methods;
-  const { fields, append, remove } = useFieldArray({
-    name: 'contents',
-    control,
-  });
-
-  /** 타입에 따른 Contents 블럭 포멧 지정 함수 */
-  const itemDataFormatByType = (type: NoticeContentsType) => {
-    const data: ItemsType = {
-      order: 0,
-      type,
-    };
-
-    switch (type) {
-      case 'body':
-      case 'subtitle':
-      case 'note':
-        data.description = '';
-        break;
-      case 'button':
-        data.buttonName = '';
-        data.buttonLink = '';
-        break;
-      case 'image':
-        data.imageUrl = '';
-      default:
-        data;
-    }
-    return data;
-  };
-
-  /** Contents 블럭 추가 함수 */
-  const handleAddContents = (type: NoticeContentsType) => () => {
-    append(itemDataFormatByType(type));
-  };
-
-  /** Contents 블럭 삭제 함수 */
-  const handleDeleteBlock = (order: number) => {
-    remove(order);
-  };
-
-  /** 공지 작성 데이터 포맷 */
-  const formatNoticeData = (originData: NoticeCreateType) => {
-    // order 정리 및 이미지 url 초기화
-    const updatedContents = originData.contents.map((item, index) => {
-      const newContents = { ...item, order: index + 1 };
-      if (newContents.type === 'image') {
-        newContents.imageUrl = '';
-      }
-      return newContents;
-    });
-
-    // 새로운 데이터 객체 반환
-    const noticeData: NoticeCreateType = {
-      ...originData,
-      contents: updatedContents,
-    };
-    return noticeData;
-  };
-
-  /** 이미지 업로드 데이터 포맷 */
-  const formatImageData = (originData: NoticeCreateType) => {
-    const imageExtensionData = originData.contents
-      .map((item, index) => {
-        return { ...item, order: index + 1 };
-      })
-      .filter((item) => item.type === 'image' && item.imageUrl !== '')
-      .map(({ order, imageUrl }) => {
-        return {
-          order: order,
-          extension: (imageUrl as File).type.split('/')[1],
-        };
-      });
-
-    const imageFileData = originData.contents
-      .filter((item) => item.type === 'image' && item.imageUrl !== '')
-      .map((item) => {
-        return item.imageUrl as File;
-      });
-
-    return { imageExtensionData, imageFileData };
-  };
+  const { register, handleSubmit, formState: { errors, isValid }} = methods;
 
   const uploadImageMutation = useMutation({
     mutationFn: uploadNoticeImages,
@@ -170,7 +121,7 @@ export default function CreateNotice() {
           <div className={styles.field}>
             <input
               className={styles.rowInput}
-              placeholder="제목 겸 알림 메시지 문구를 입력해 주세요. (최대 30자)"
+              placeholder="제목 또는 알림 메시지 문구를 입력해 주세요. (최대 30자)"
               {...register('title', noticeTitleRules)}
             />
             <p className={styles.rowErrorMessage}>{errors.title && errors.title?.message}</p>
@@ -187,21 +138,7 @@ export default function CreateNotice() {
             <p className={styles.rowErrorMessage}>{errors.description && errors.description.message}</p>
           </div>
         </div>
-        <section>
-          {fields.map((field, index) => (
-            <ContentsContainer key={field.id} content={field} order={index} handleDeleteBlock={handleDeleteBlock} />
-          ))}
-        </section>
-        <section className={styles.contents}>
-          {Object.entries(NOTICE_CONTENT).map(([key, value], index) => (
-            <button
-              key={index}
-              className={styles.block}
-              onClick={handleAddContents(key as NoticeContentsType)}
-              type="button"
-            >{`+ ${value} 추가`}</button>
-          ))}
-        </section>
+        <ContentsBody />
         <button
           type="button"
           onClick={handleSubmit(onSubmit)}
